week1/requirements.txt
acme==3.2.0
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
aiosmtplib==3.0.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
bcrypt==3.2.2
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
certbot==3.2.0
certbot-nginx==3.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7.1
configobj==5.0.9
cryptography==44.0.3
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-mail==1.5.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fqdn==1.5.1
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jose==1.0.0
josepy==1.15.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
multidict==6.4.4
mysql-connector-python==9.3.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.4
passlib==1.7.4
pefile==2023.2.7
platformdirs==4.3.6
prettytable==3.14.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.1
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-jose==3.4.0
python-json-logger==3.2.1
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.6
rpds-py==0.22.3
rsa==4.9.1
Send2Trash==1.8.3
setuptools==75.6.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typer==0.15.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
yarl==1.20.0
